+++ packages/middleware-eventstream/src/eventStreamConfiguration.ts
