+++ packages/middleware-eventstream/src/eventStreamHeaderMiddleware.ts
