+++ packages/middleware-eventstream/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./configuration";
-export * from "./handling-middleware";
-export * from "./headers-middleware";
-export * from "./plugin";
+export * from "./eventStreamConfiguration";
+export * from "./eventStreamHandlingMiddleware";
+export * from "./eventStreamHeaderMiddleware";
+export * from "./getEventStreamPlugin";
