+++ packages/middleware-sdk-s3-control/src/process-arnables-plugin/getProcessArnablesPlugin.ts
