+++ packages/middleware-sdk-s3/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./check-content-length-header";
-export * from "./configuration";
+export * from "./s3Configuration";
 export * from "./throw-200-exceptions";
 export * from "./validate-bucket-name";
