+++ packages/middleware-sdk-s3/src/s3Configuration.ts
