+++ packages/middleware-sdk-transcribe-streaming/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./eventstream-handler";
+export * from "./getTranscribeStreamingPlugin";
 export * from "./middleware-inject-response-values";
 export * from "./middleware-port";
-export * from "./plugin";
