+++ packages/middleware-signing/src/awsAuthConfiguration.ts
